add Base model and inherit in Todo
